feat(server): check if user can delete

+++ server/src/mutation.ts
@@ -46,13 +46,25 @@ const Mutation = {
   },
   async deleteItem(parent: any, args: any, context: Context, info: any) {
     const { id } = args;
-    const item = await context.db.mutation.deleteItem(
+    const { user, userId } = context.request;
+    const item = await context.db.query.item(
+      { where: { id } },
+      `{ user { id } }`
+    );
+    const ownsItems = item.user.id === userId;
+    const hasPermissions = user.permissions.some(permission =>
+      ['ADMIN', 'ITEM_DELETE'].includes(permission)
+    );
+    if (!ownsItems || !hasPermissions) {
+      throw new Error("you can't do that");
+    }
+    const itemToDelete = await context.db.mutation.deleteItem(
       {
         where: { id }
       },
       info
     );
-    return item;
+    return itemToDelete;
   },
   async signUp(parent: any, args: any, context: Context, info: any) {
     const { name, email, password } = args;
